@@ -1,89 +1,148 @@
 {
+  "_args": [
+    [
+      {
+        "raw": "cordova-common@^2.1.0",
+        "scope": null,
+        "escapedName": "cordova-common",
+        "name": "cordova-common",
+        "rawSpec": "^2.1.0",
+        "spec": ">=2.1.0 <3.0.0",
+        "type": "range"
+      },
+      "/Users/jbowser/cordova/cordova-android"
+    ]
+  ],
+  "_from": "cordova-common@>=2.1.0 <3.0.0",
+  "_id": "cordova-common@2.1.0",
+  "_inBundle": true,
+  "_inCache": true,
+  "_location": "/cordova-android/cordova-common",
+  "_nodeVersion": "6.6.0",
+  "_npmOperationalInternal": {
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/cordova-common-2.1.0.tgz_1504306462859_0.04064799007028341"
+  },
+  "_npmUser": {
+    "name": "audreyso",
+    "email": "audreyeso@gmail.com"
+  },
+  "_npmVersion": "4.6.1",
+  "_phantomChildren": {},
+  "_requested": {
+    "raw": "cordova-common@^2.1.0",
+    "scope": null,
+    "escapedName": "cordova-common",
+    "name": "cordova-common",
+    "rawSpec": "^2.1.0",
+    "spec": ">=2.1.0 <3.0.0",
+    "type": "range"
+  },
+  "_requiredBy": [
+    "/cordova-android"
+  ],
+  "_resolved": "https://registry.npmjs.org/cordova-common/-/cordova-common-2.1.0.tgz",
+  "_shasum": "bb357ee1b9825031ed9db3c56b592efe973d1640",
+  "_shrinkwrap": null,
+  "_spec": "cordova-common@^2.1.0",
+  "_where": "/Users/jbowser/cordova/cordova-android",
   "author": {
     "name": "Apache Software Foundation"
-  },
-  "name": "cordova-common",
-  "description": "Apache Cordova tools and platforms shared routines",
-  "license": "Apache-2.0",
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": "git://git-wip-us.apache.org/repos/asf/cordova-common.git"
   },
   "bugs": {
     "url": "https://issues.apache.org/jira/browse/CB",
     "email": "dev@cordova.apache.org"
   },
-  "main": "cordova-common.js",
-  "engines": {
-    "node": ">=0.9.9"
-  },
-  "scripts": {
-    "test": "npm run jshint && npm run jasmine",
-    "jshint": "node node_modules/jshint/bin/jshint src && node node_modules/jshint/bin/jshint spec",
-    "jasmine": "node node_modules/jasmine-node/bin/jasmine-node --captureExceptions --color spec",
-    "cover": "node node_modules/istanbul/lib/cli.js cover --root src --print detail node_modules/jasmine-node/bin/jasmine-node -- spec"
-  },
-  "engineStrict": true,
+  "contributors": [],
   "dependencies": {
+    "ansi": "^0.3.1",
     "bplist-parser": "^0.1.0",
     "cordova-registry-mapper": "^1.1.8",
-    "elementtree": "^0.1.6",
+    "elementtree": "0.1.6",
     "glob": "^5.0.13",
+    "minimatch": "^3.0.0",
     "osenv": "^0.1.3",
-    "plist": "^1.1.0",
+    "plist": "^1.2.0",
     "q": "^1.4.1",
     "semver": "^5.0.1",
-    "shelljs": "^0.5.1",
+    "shelljs": "^0.5.3",
     "underscore": "^1.8.3",
     "unorm": "^1.3.3"
   },
+  "description": "Apache Cordova tools and platforms shared routines",
   "devDependencies": {
-    "istanbul": "^0.3.17",
-    "jasmine-node": "^1.14.5",
-    "jshint": "^2.8.0"
+    "eslint": "^4.0.0",
+    "eslint-config-semistandard": "^11.0.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-node": "^5.0.0",
+    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "istanbul": "^0.4.5",
+    "jasmine": "^2.5.2",
+    "promise-matchers": "^0.9.6",
+    "rewire": "^2.5.1"
   },
-  "contributors": [],
-  "_id": "cordova-common@1.0.0",
-  "_shasum": "b21947e89a4a89292ec563abf9ee6ccb2b9f3aef",
-  "_resolved": "http://registry.npmjs.org/cordova-common/-/cordova-common-1.0.0.tgz",
-  "_from": "cordova-common@>=1.0.0 <1.1.0",
-  "_npmVersion": "2.14.7",
-  "_nodeVersion": "4.2.1",
-  "_npmUser": {
-    "name": "kotikov.vladimir",
-    "email": "kotikov.vladimir@gmail.com"
-  },
+  "directories": {},
   "dist": {
-    "shasum": "b21947e89a4a89292ec563abf9ee6ccb2b9f3aef",
-    "tarball": "http://registry.npmjs.org/cordova-common/-/cordova-common-1.0.0.tgz"
+    "shasum": "bb357ee1b9825031ed9db3c56b592efe973d1640",
+    "tarball": "https://registry.npmjs.org/cordova-common/-/cordova-common-2.1.0.tgz"
   },
+  "engines": {
+    "node": ">=4.0.0"
+  },
+  "homepage": "https://github.com/apache/cordova-lib#readme",
+  "license": "Apache-2.0",
+  "main": "cordova-common.js",
   "maintainers": [
     {
-      "name": "bowserj",
-      "email": "bowserj@apache.org"
+      "name": "audreyso",
+      "email": "audreyeso@gmail.com"
     },
     {
-      "name": "kotikov.vladimir",
-      "email": "kotikov.vladimir@gmail.com"
+      "name": "apachebuilds",
+      "email": "root@apache.org"
     },
     {
-      "name": "purplecabbage",
-      "email": "purplecabbage@gmail.com"
+      "name": "filmaj",
+      "email": "maj.fil@gmail.com"
+    },
+    {
+      "name": "timbarham",
+      "email": "npmjs@barhams.info"
     },
     {
       "name": "shazron",
       "email": "shazron@gmail.com"
     },
     {
+      "name": "bowserj",
+      "email": "bowserj@apache.org"
+    },
+    {
+      "name": "purplecabbage",
+      "email": "purplecabbage@gmail.com"
+    },
+    {
       "name": "stevegill",
       "email": "stevengill97@gmail.com"
     },
     {
-      "name": "timbarham",
-      "email": "npmjs@barhams.info"
+      "name": "kotikov.vladimir",
+      "email": "kotikov.vladimir@gmail.com"
     }
   ],
-  "directories": {},
-  "readme": "ERROR: No README data found!"
+  "name": "cordova-common",
+  "optionalDependencies": {},
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/apache/cordova-lib.git"
+  },
+  "scripts": {
+    "cover": "istanbul cover --root src --print detail jasmine",
+    "eslint": "eslint src && eslint spec",
+    "jasmine": "jasmine JASMINE_CONFIG_PATH=spec/support/jasmine.json",
+    "test": "npm run eslint && npm run jasmine"
+  },
+  "version": "2.1.0"
 }
