@@ -25,11 +25,11 @@ var CordovaError = require('cordova-common').CordovaError;
 
 var Adb = {};
 
-function isDevice(line) {
+function isDevice (line) {
     return line.match(/\w+\tdevice/) && !line.match(/emulator/);
 }
 
-function isEmulator(line) {
+function isEmulator (line) {
     return line.match(/device/) && line.match(/emulator/);
 }
 
@@ -44,8 +44,7 @@ function isEmulator(line) {
  *   devices/emulators
  */
 Adb.devices = function (opts) {
-    return spawn('adb', ['devices'], {cwd: os.tmpdir()})
-    .then(function(output) {
+    return spawn('adb', ['devices'], {cwd: os.tmpdir()}).then(function (output) {
         return output.split('\n').filter(function (line) {
             // Filter out either real devices or emulators, depending on options
             return (line && opts && opts.emulators) ? isEmulator(line) : isDevice(line);
@@ -56,38 +55,44 @@ Adb.devices = function (opts) {
 };
 
 Adb.install = function (target, packagePath, opts) {
-    events.emit('verbose', 'Installing apk ' + packagePath + ' on ' + target + '...');
+    events.emit('verbose', 'Installing apk ' + packagePath + ' on target ' + target + '...');
     var args = ['-s', target, 'install'];
     if (opts && opts.replace) args.push('-r');
-    return spawn('adb', args.concat(packagePath), {cwd: os.tmpdir()})
-    .then(function(output) {
+    return spawn('adb', args.concat(packagePath), {cwd: os.tmpdir()}).then(function (output) {
         // 'adb install' seems to always returns no error, even if installation fails
         // so we catching output to detect installation failure
-        if (output.match(/Failure/))
+        if (output.match(/Failure/)) {
+            if (output.match(/INSTALL_PARSE_FAILED_NO_CERTIFICATES/)) {
+                output += '\n\n' + 'Sign the build using \'-- --keystore\' or \'--buildConfig\'' +
+                    ' or sign and deploy the unsigned apk manually using Android tools.';
+            } else if (output.match(/INSTALL_FAILED_VERSION_DOWNGRADE/)) {
+                output += '\n\n' + 'You\'re trying to install apk with a lower versionCode that is already installed.' +
+                    '\nEither uninstall an app or increment the versionCode.';
+            }
+
             return Q.reject(new CordovaError('Failed to install apk to device: ' + output));
+        }
     });
 };
 
 Adb.uninstall = function (target, packageId) {
-    events.emit('verbose', 'Uninstalling ' + packageId + ' from ' + target + '...');
+    events.emit('verbose', 'Uninstalling package ' + packageId + ' from target ' + target + '...');
     return spawn('adb', ['-s', target, 'uninstall', packageId], {cwd: os.tmpdir()});
 };
 
 Adb.shell = function (target, shellCommand) {
-    events.emit('verbose', 'Running command "' + shellCommand + '" on ' + target + '...');
+    events.emit('verbose', 'Running adb shell command "' + shellCommand + '" on target ' + target + '...');
     var args = ['-s', target, 'shell'];
     shellCommand = shellCommand.split(/\s+/);
-    return spawn('adb', args.concat(shellCommand), {cwd: os.tmpdir()})
-    .catch(function (output) {
+    return spawn('adb', args.concat(shellCommand), {cwd: os.tmpdir()}).catch(function (output) {
         return Q.reject(new CordovaError('Failed to execute shell command "' +
             shellCommand + '"" on device: ' + output));
     });
 };
 
 Adb.start = function (target, activityName) {
-    events.emit('verbose', 'Starting application "' + activityName + '" on ' + target + '...');
-    return Adb.shell(target, 'am start -W -a android.intent.action.MAIN -n' + activityName)
-    .catch(function (output) {
+    events.emit('verbose', 'Starting application "' + activityName + '" on target ' + target + '...');
+    return Adb.shell(target, 'am start -W -a android.intent.action.MAIN -n' + activityName).catch(function (output) {
         return Q.reject(new CordovaError('Failed to start application "' +
             activityName + '"" on device: ' + output));
     });
