@@ -19,10 +19,10 @@
        under the License.
 */
 
-var Q       = require('q'),
-    path    = require('path'),
-    fs      = require('fs'),
-    nopt = require('nopt');
+var Q = require('q');
+var path = require('path');
+var fs = require('fs');
+var nopt = require('nopt');
 
 var Adb = require('./Adb');
 
@@ -31,7 +31,7 @@ var events = require('cordova-common').events;
 var spawn = require('cordova-common').superspawn.spawn;
 var CordovaError = require('cordova-common').CordovaError;
 
-function parseOpts(options, resolvedTarget) {
+function parseOpts (options, resolvedTarget, projectRoot) {
     options = options || {};
     options.argv = nopt({
         gradle: Boolean,
@@ -39,7 +39,7 @@ function parseOpts(options, resolvedTarget) {
         prepenv: Boolean,
         versionCode: String,
         minSdkVersion: String,
-        gradleArg: String,
+        gradleArg: [String, Array],
         keystore: path,
         alias: String,
         storePassword: String,
@@ -55,28 +55,24 @@ function parseOpts(options, resolvedTarget) {
         extraArgs: []
     };
 
-    if (options.argv.ant || options.argv.gradle)
-        ret.buildMethod = options.argv.ant ? 'ant' : 'gradle';
+    if (options.argv.ant || options.argv.gradle) { ret.buildMethod = options.argv.ant ? 'ant' : 'gradle'; }
 
     if (options.nobuild) ret.buildMethod = 'none';
 
-    if (options.argv.versionCode)
-        ret.extraArgs.push('-PcdvVersionCode=' + options.argv.versionCode);
+    if (options.argv.versionCode) { ret.extraArgs.push('-PcdvVersionCode=' + options.argv.versionCode); }
 
-    if (options.argv.minSdkVersion)
-        ret.extraArgs.push('-PcdvMinSdkVersion=' + options.argv.minSdkVersion);
+    if (options.argv.minSdkVersion) { ret.extraArgs.push('-PcdvMinSdkVersion=' + options.argv.minSdkVersion); }
 
-    if (options.argv.gradleArg)
-        ret.extraArgs.push(options.argv.gradleArg);
+    if (options.argv.gradleArg) {
+        ret.extraArgs = ret.extraArgs.concat(options.argv.gradleArg);
+    }
 
     var packageArgs = {};
 
-    if (options.argv.keystore)
-        packageArgs.keystore = path.relative(this.root, path.resolve(options.argv.keystore));
+    if (options.argv.keystore) { packageArgs.keystore = path.relative(projectRoot, path.resolve(options.argv.keystore)); }
 
-    ['alias','storePassword','password','keystoreType'].forEach(function (flagName) {
-        if (options.argv[flagName])
-            packageArgs[flagName] = options.argv[flagName];
+    ['alias', 'storePassword', 'password', 'keystoreType'].forEach(function (flagName) {
+        if (options.argv[flagName]) { packageArgs[flagName] = options.argv[flagName]; }
     });
 
     var buildConfig = options.buildConfig;
@@ -87,21 +83,20 @@ function parseOpts(options, resolvedTarget) {
         if (!fs.existsSync(buildConfig)) {
             throw new Error('Specified build config file does not exist: ' + buildConfig);
         }
-        events.emit('log', 'Reading build config file: '+ path.resolve(buildConfig));
+        events.emit('log', 'Reading build config file: ' + path.resolve(buildConfig));
         var buildjson = fs.readFileSync(buildConfig, 'utf8');
-        //var config = JSON.parse(fs.readFileSync(buildConfig, 'utf8'));
-        var config = JSON.parse(buildjson);
+        var config = JSON.parse(buildjson.replace(/^\ufeff/, '')); // Remove BOM
         if (config.android && config.android[ret.buildType]) {
             var androidInfo = config.android[ret.buildType];
-            if(androidInfo.keystore && !packageArgs.keystore) {
-                if(androidInfo.keystore.substr(0,1) === '~') {
+            if (androidInfo.keystore && !packageArgs.keystore) {
+                if (androidInfo.keystore.substr(0, 1) === '~') {
                     androidInfo.keystore = process.env.HOME + androidInfo.keystore.substr(1);
                 }
                 packageArgs.keystore = path.resolve(path.dirname(buildConfig), androidInfo.keystore);
                 events.emit('log', 'Reading the keystore from: ' + packageArgs.keystore);
             }
 
-            ['alias', 'storePassword', 'password','keystoreType'].forEach(function (key){
+            ['alias', 'storePassword', 'password', 'keystoreType'].forEach(function (key) {
                 packageArgs[key] = packageArgs[key] || androidInfo[key];
             });
         }
@@ -112,8 +107,8 @@ function parseOpts(options, resolvedTarget) {
             packageArgs.password, packageArgs.keystoreType);
     }
 
-    if(!ret.packageInfo) {
-        if(Object.keys(packageArgs).length > 0) {
+    if (!ret.packageInfo) {
+        if (Object.keys(packageArgs).length > 0) {
             events.emit('warn', '\'keystore\' and \'alias\' need to be specified to generate a signed archive.');
         }
     }
@@ -125,11 +120,10 @@ function parseOpts(options, resolvedTarget) {
  * Builds the project with the specifed options
  * Returns a promise.
  */
-module.exports.runClean = function(options) {
-    var opts = parseOpts(options);
+module.exports.runClean = function (options) {
+    var opts = parseOpts(options, null, this.root);
     var builder = builders.getBuilder(opts.buildMethod);
-    return builder.prepEnv(opts)
-    .then(function() {
+    return builder.prepEnv(opts).then(function () {
         return builder.clean(opts);
     });
 };
@@ -146,20 +140,17 @@ module.exports.runClean = function(options) {
  * @return  {Promise<Object>}            Promise, resolved with built packages
  *   information.
  */
-module.exports.run = function(options, optResolvedTarget) {
-    var opts = parseOpts(options, optResolvedTarget);
+module.exports.run = function (options, optResolvedTarget) {
+    var opts = parseOpts(options, optResolvedTarget, this.root);
     var builder = builders.getBuilder(opts.buildMethod);
-    var self = this;
-    return builder.prepEnv(opts)
-    .then(function() {
+    return builder.prepEnv(opts).then(function () {
         if (opts.prepEnv) {
-            self.events.emit('verbose', 'Build file successfully prepared.');
+            events.emit('verbose', 'Build file successfully prepared.');
             return;
         }
-        return builder.build(opts)
-        .then(function() {
+        return builder.build(opts).then(function () {
             var apkPaths = builder.findOutputApks(opts.buildType, opts.arch);
-            self.events.emit('log', 'Built the following apk(s): \n\t' + apkPaths.join('\n\t'));
+            events.emit('log', 'Built the following apk(s): \n\t' + apkPaths.join('\n\t'));
             return {
                 apkPaths: apkPaths,
                 buildType: opts.buildType,
@@ -169,47 +160,35 @@ module.exports.run = function(options, optResolvedTarget) {
     });
 };
 
-// Called by plugman after installing plugins, and by create script after creating project.
-module.exports.prepBuildFiles = function() {
-    return builders.getBuilder('gradle').prepBuildFiles();
-};
-
 /*
  * Detects the architecture of a device/emulator
  * Returns "arm" or "x86".
  */
-module.exports.detectArchitecture = function(target) {
-    function helper() {
-        return Adb.shell(target, 'cat /proc/cpuinfo')
-        .then(function(output) {
+module.exports.detectArchitecture = function (target) {
+    function helper () {
+        return Adb.shell(target, 'cat /proc/cpuinfo').then(function (output) {
             return /intel/i.exec(output) ? 'x86' : 'arm';
         });
     }
     // It sometimes happens (at least on OS X), that this command will hang forever.
     // To fix it, either unplug & replug device, or restart adb server.
-    return helper()
-    .timeout(1000, new CordovaError('Device communication timed out. Try unplugging & replugging the device.'))
-    .then(null, function(err) {
+    return helper().timeout(1000, new CordovaError('Device communication timed out. Try unplugging & replugging the device.')).then(null, function (err) {
         if (/timed out/.exec('' + err)) {
             // adb kill-server doesn't seem to do the trick.
             // Could probably find a x-platform version of killall, but I'm not actually
             // sure that this scenario even happens on non-OSX machines.
-            return spawn('killall', ['adb'])
-            .then(function() {
-                events.emit('verbose', 'adb seems hung. retrying.');
-                return helper()
-                .then(null, function() {
+            events.emit('verbose', 'adb timed out while detecting device/emulator architecture. Killing adb and trying again.');
+            return spawn('killall', ['adb']).then(function () {
+                return helper().then(null, function () {
                     // The double kill is sadly often necessary, at least on mac.
-                    events.emit('warn', 'Now device not found... restarting adb again.');
-                    return spawn('killall', ['adb'])
-                    .then(function() {
-                        return helper()
-                        .then(null, function() {
-                            return Q.reject(new CordovaError('USB is flakey. Try unplugging & replugging the device.'));
+                    events.emit('warn', 'adb timed out a second time while detecting device/emulator architecture. Killing adb and trying again.');
+                    return spawn('killall', ['adb']).then(function () {
+                        return helper().then(null, function () {
+                            return Q.reject(new CordovaError('adb timed out a third time while detecting device/emulator architecture. Try unplugging & replugging the device.'));
                         });
                     });
                 });
-            }, function() {
+            }, function () {
                 // For non-killall OS's.
                 return Q.reject(err);
             });
@@ -218,10 +197,10 @@ module.exports.detectArchitecture = function(target) {
     });
 };
 
-module.exports.findBestApkForArchitecture = function(buildResults, arch) {
-    var paths = buildResults.apkPaths.filter(function(p) {
+module.exports.findBestApkForArchitecture = function (buildResults, arch) {
+    var paths = buildResults.apkPaths.filter(function (p) {
         var apkName = path.basename(p);
-        if (buildResults.buildType == 'debug') {
+        if (buildResults.buildType === 'debug') {
             return /-debug/.exec(apkName);
         }
         return !/-debug/.exec(apkName);
@@ -241,7 +220,7 @@ module.exports.findBestApkForArchitecture = function(buildResults, arch) {
     throw new Error('Could not find apk architecture: ' + arch + ' build-type: ' + buildResults.buildType);
 };
 
-function PackageInfo(keystore, alias, storePassword, password, keystoreType) {
+function PackageInfo (keystore, alias, storePassword, password, keystoreType) {
     this.keystore = {
         'name': 'key.store',
         'value': keystore
@@ -271,10 +250,10 @@ function PackageInfo(keystore, alias, storePassword, password, keystoreType) {
 }
 
 PackageInfo.prototype = {
-    toProperties: function() {
+    toProperties: function () {
         var self = this;
         var result = '';
-        Object.keys(self).forEach(function(key) {
+        Object.keys(self).forEach(function (key) {
             result += self[key].name;
             result += '=';
             result += self[key].value.replace(/\\/g, '\\\\');
@@ -284,7 +263,7 @@ PackageInfo.prototype = {
     }
 };
 
-module.exports.help = function() {
+module.exports.help = function () {
     console.log('Usage: ' + path.relative(process.cwd(), path.join('../build')) + ' [flags] [Signed APK flags]');
     console.log('Flags:');
     console.log('    \'--debug\': will build project in debug mode (default)');
