@@ -21,40 +21,42 @@
 
 /* jshint sub:true */
 
-var retry      = require('./retry');
-var build      = require('./build');
-var check_reqs = require('./check_reqs');
+var android_versions = require('android-versions');
+var retry = require('./retry');
+var build = require('./build');
 var path = require('path');
 var Adb = require('./Adb');
 var AndroidManifest = require('./AndroidManifest');
 var events = require('cordova-common').events;
-var spawn = require('cordova-common').superspawn.spawn;
+var superspawn = require('cordova-common').superspawn;
 var CordovaError = require('cordova-common').CordovaError;
+var shelljs = require('shelljs');
+var android_sdk = require('./android_sdk');
+var check_reqs = require('./check_reqs');
 
-var Q             = require('q');
-var os            = require('os');
+var Q = require('q');
+var os = require('os');
+var fs = require('fs');
 var child_process = require('child_process');
 
 // constants
-var ONE_SECOND              = 1000; // in milliseconds
-var ONE_MINUTE              = 60 * ONE_SECOND; // in milliseconds
+var ONE_SECOND = 1000; // in milliseconds
+var ONE_MINUTE = 60 * ONE_SECOND; // in milliseconds
 var INSTALL_COMMAND_TIMEOUT = 5 * ONE_MINUTE; // in milliseconds
-var NUM_INSTALL_RETRIES     = 3;
-var EXEC_KILL_SIGNAL        = 'SIGKILL';
+var NUM_INSTALL_RETRIES = 3;
+var CHECK_BOOTED_INTERVAL = 3 * ONE_SECOND; // in milliseconds
+var EXEC_KILL_SIGNAL = 'SIGKILL';
 
-/**
- * Returns a Promise for a list of emulator images in the form of objects
- * {
-       name   : <emulator_name>,
-       path   : <path_to_emulator_image>,
-       target : <api_target>,
-       abi    : <cpu>,
-       skin   : <skin>
-   }
- */
-module.exports.list_images = function() {
-    return spawn('android', ['list', 'avds'])
-    .then(function(output) {
+function forgivingWhichSync (cmd) {
+    try {
+        return fs.realpathSync(shelljs.which(cmd));
+    } catch (e) {
+        return '';
+    }
+}
+
+module.exports.list_images_using_avdmanager = function () {
+    return superspawn.spawn('avdmanager', ['list', 'avd']).then(function (output) {
         var response = output.split('\n');
         var emulator_list = [];
         for (var i = 1; i < response.length; i++) {
@@ -62,13 +64,78 @@ module.exports.list_images = function() {
             var img_obj = {};
             if (response[i].match(/Name:\s/)) {
                 img_obj['name'] = response[i].split('Name: ')[1].replace('\r', '');
+                if (response[i + 1].match(/Device:\s/)) {
+                    i++;
+                    img_obj['device'] = response[i].split('Device: ')[1].replace('\r', '');
+                }
                 if (response[i + 1].match(/Path:\s/)) {
                     i++;
                     img_obj['path'] = response[i].split('Path: ')[1].replace('\r', '');
                 }
-                if (response[i + 1].match(/\(API\slevel\s/)) {
+                if (response[i + 1].match(/Target:\s/)) {
                     i++;
-                    img_obj['target'] = response[i].replace('\r', '');
+                    if (response[i + 1].match(/ABI:\s/)) {
+                        img_obj['abi'] = response[i + 1].split('ABI: ')[1].replace('\r', '');
+                    }
+                    // This next conditional just aims to match the old output of `android list avd`
+                    // We do so so that we don't have to change the logic when parsing for the
+                    // best emulator target to spawn (see below in `best_image`)
+                    // This allows us to transitionally support both `android` and `avdmanager` binaries,
+                    // depending on what SDK version the user has
+                    if (response[i + 1].match(/Based\son:\s/)) {
+                        img_obj['target'] = response[i + 1].split('Based on:')[1];
+                        if (img_obj['target'].match(/Tag\/ABI:\s/)) {
+                            img_obj['target'] = img_obj['target'].split('Tag/ABI:')[0].replace('\r', '').trim();
+                            if (img_obj['target'].indexOf('(') > -1) {
+                                img_obj['target'] = img_obj['target'].substr(0, img_obj['target'].indexOf('(') - 1).trim();
+                            }
+                        }
+                        var version_string = img_obj['target'].replace(/Android\s+/, '');
+
+                        var api_level = android_sdk.version_string_to_api_level[version_string];
+                        if (api_level) {
+                            img_obj['target'] += ' (API level ' + api_level + ')';
+                        }
+                    }
+                }
+                if (response[i + 1].match(/Skin:\s/)) {
+                    i++;
+                    img_obj['skin'] = response[i].split('Skin: ')[1].replace('\r', '');
+                }
+
+                emulator_list.push(img_obj);
+            }
+            /* To just return a list of names use this
+            if (response[i].match(/Name:\s/)) {
+                emulator_list.push(response[i].split('Name: ')[1].replace('\r', '');
+            } */
+
+        }
+        return emulator_list;
+    });
+};
+
+module.exports.list_images_using_android = function () {
+    return superspawn.spawn('android', ['list', 'avd']).then(function (output) {
+        var response = output.split('\n');
+        var emulator_list = [];
+        for (var i = 1; i < response.length; i++) {
+            // To return more detailed information use img_obj
+            var img_obj = {};
+            if (response[i].match(/Name:\s/)) {
+                img_obj['name'] = response[i].split('Name: ')[1].replace('\r', '');
+                if (response[i + 1].match(/Device:\s/)) {
+                    i++;
+                    img_obj['device'] = response[i].split('Device: ')[1].replace('\r', '');
+                }
+                if (response[i + 1].match(/Path:\s/)) {
+                    i++;
+                    img_obj['path'] = response[i].split('Path: ')[1].replace('\r', '');
+                }
+                if (response[i + 1].match(/\(API\slevel\s/) || (response[i + 2] && response[i + 2].match(/\(API\slevel\s/))) {
+                    i++;
+                    var secondLine = response[i + 1].match(/\(API\slevel\s/) ? response[i + 1] : '';
+                    img_obj['target'] = (response[i] + secondLine).split('Target: ')[1].replace('\r', '');
                 }
                 if (response[i + 1].match(/ABI:\s/)) {
                     i++;
@@ -84,10 +151,47 @@ module.exports.list_images = function() {
             /* To just return a list of names use this
             if (response[i].match(/Name:\s/)) {
                 emulator_list.push(response[i].split('Name: ')[1].replace('\r', '');
-            }*/
+            } */
 
         }
         return emulator_list;
+    });
+};
+
+/**
+ * Returns a Promise for a list of emulator images in the form of objects
+ * {
+       name   : <emulator_name>,
+       device : <device>,
+       path   : <path_to_emulator_image>,
+       target : <api_target>,
+       abi    : <cpu>,
+       skin   : <skin>
+   }
+ */
+module.exports.list_images = function () {
+    return Q.fcall(function () {
+        if (forgivingWhichSync('avdmanager')) {
+            return module.exports.list_images_using_avdmanager();
+        } else if (forgivingWhichSync('android')) {
+            return module.exports.list_images_using_android();
+        } else {
+            return Q().then(function () {
+                throw new CordovaError('Could not find either `android` or `avdmanager` on your $PATH! Are you sure the Android SDK is installed and available?');
+            });
+        }
+    }).then(function (avds) {
+        // In case we're missing the Android OS version string from the target description, add it.
+        return avds.map(function (avd) {
+            if (avd.target && avd.target.indexOf('Android API') > -1 && avd.target.indexOf('API level') < 0) {
+                var api_level = avd.target.match(/\d+/);
+                if (api_level) {
+                    var level = android_versions.get(api_level);
+                    avd.target = 'Android ' + level.semver + ' (API level ' + api_level + ')';
+                }
+            }
+            return avd;
+        });
     });
 };
 
@@ -96,20 +200,19 @@ module.exports.list_images = function() {
  * or undefined if no avds exist.
  * Returns a promise.
  */
-module.exports.best_image = function() {
-    return this.list_images()
-    .then(function(images) {
+module.exports.best_image = function () {
+    return this.list_images().then(function (images) {
         // Just return undefined if there is no images
         if (images.length === 0) return;
 
         var closest = 9999;
         var best = images[0];
-        var project_target = check_reqs.get_target().replace('android-', '');
+        var project_target = parseInt(check_reqs.get_target().replace('android-', ''));
         for (var i in images) {
             var target = images[i].target;
-            if(target) {
-                var num = target.split('(API level ')[1].replace(')', '');
-                if (num == project_target) {
+            if (target && target.indexOf('API level') > -1) {
+                var num = parseInt(target.split('(API level ')[1].replace(')', ''));
+                if (num === project_target) {
                     return images[i];
                 } else if (project_target - num < closest && project_target > num) {
                     closest = project_target - num;
@@ -122,18 +225,18 @@ module.exports.best_image = function() {
 };
 
 // Returns a promise.
-module.exports.list_started = function() {
+module.exports.list_started = function () {
     return Adb.devices({emulators: true});
 };
 
 // Returns a promise.
-module.exports.list_targets = function() {
-    return spawn('android', ['list', 'targets'], {cwd: os.tmpdir()})
-    .then(function(output) {
+// TODO: we should remove this, there's a more robust method under android_sdk.js
+module.exports.list_targets = function () {
+    return superspawn.spawn('android', ['list', 'targets'], {cwd: os.tmpdir()}).then(function (output) {
         var target_out = output.split('\n');
         var targets = [];
         for (var i = target_out.length; i >= 0; i--) {
-            if(target_out[i].match(/id:/)) {
+            if (target_out[i].match(/id:/)) {
                 targets.push(targets[i].split(' ')[1]);
             }
         }
@@ -142,21 +245,40 @@ module.exports.list_targets = function() {
 };
 
 /*
+ * Gets unused port for android emulator, between 5554 and 5584
+ * Returns a promise.
+ */
+module.exports.get_available_port = function () {
+    var self = this;
+
+    return self.list_started().then(function (emulators) {
+        for (var p = 5584; p >= 5554; p -= 2) {
+            if (emulators.indexOf('emulator-' + p) === -1) {
+                events.emit('verbose', 'Found available port: ' + p);
+                return p;
+            }
+        }
+        throw new CordovaError('Could not find an available avd port');
+    });
+};
+
+/*
  * Starts an emulator with the given ID,
  * and returns the started ID of that emulator.
  * If no ID is given it will use the first image available,
  * if no image is available it will error out (maybe create one?).
+ * If no boot timeout is given or the value is negative it will wait forever for
+ * the emulator to boot
  *
  * Returns a promise.
  */
-module.exports.start = function(emulator_ID) {
+module.exports.start = function (emulator_ID, boot_timeout) {
     var self = this;
 
-    return Q().then(function() {
+    return Q().then(function () {
         if (emulator_ID) return Q(emulator_ID);
 
-        return self.best_image()
-        .then(function(best) {
+        return self.best_image().then(function (best) {
             if (best && best.name) {
                 events.emit('warn', 'No emulator specified, defaulting to ' + best.name);
                 return best.name;
@@ -168,77 +290,88 @@ module.exports.start = function(emulator_ID) {
                 '2. Create an AVD by running: ' + androidCmd + ' avd\n' +
                 'HINT: For a faster emulator, use an Intel System Image and install the HAXM device driver\n'));
         });
-    }).then(function(emulatorId) {
-        var uuid = 'cordova_emulator_' + new Date().getTime();
-        var uuidProp = 'emu.uuid=' + uuid;
-        var args = ['-avd', emulatorId, '-prop', uuidProp];
-        // Don't wait for it to finish, since the emulator will probably keep running for a long time.
-        child_process
-            .spawn('emulator', args, { stdio: 'inherit', detached: true })
-            .unref();
+    }).then(function (emulatorId) {
+        return self.get_available_port().then(function (port) {
+            // Figure out the directory the emulator binary runs in, and set the cwd to that directory.
+            // Workaround for https://code.google.com/p/android/issues/detail?id=235461
+            var emulator_dir = path.dirname(shelljs.which('emulator'));
+            var args = ['-avd', emulatorId, '-port', port];
+            // Don't wait for it to finish, since the emulator will probably keep running for a long time.
+            child_process
+                .spawn('emulator', args, { stdio: 'inherit', detached: true, cwd: emulator_dir })
+                .unref();
 
-        // wait for emulator to start
-        events.emit('log', 'Waiting for emulator...');
-        return self.wait_for_emulator(uuid);
-    }).then(function(emulatorId) {
-        if (!emulatorId)
-            return Q.reject(new CordovaError('Failed to start emulator'));
+            // wait for emulator to start
+            events.emit('log', 'Waiting for emulator to start...');
+            return self.wait_for_emulator(port);
+        });
+    }).then(function (emulatorId) {
+        if (!emulatorId) { return Q.reject(new CordovaError('Failed to start emulator')); }
 
-        //wait for emulator to boot up
-        process.stdout.write('Booting up emulator (this may take a while)...');
-        return self.wait_for_boot(emulatorId)
-        .then(function() {
-            events.emit('log','BOOT COMPLETE');
-            //unlock screen
-            return Adb.shell(emulatorId, 'input keyevent 82');
-        }).then(function() {
-            //return the new emulator id for the started emulators
-            return emulatorId;
+        // wait for emulator to boot up
+        process.stdout.write('Waiting for emulator to boot (this may take a while)...');
+        return self.wait_for_boot(emulatorId, boot_timeout).then(function (success) {
+            if (success) {
+                events.emit('log', 'BOOT COMPLETE');
+                // unlock screen
+                return Adb.shell(emulatorId, 'input keyevent 82').then(function () {
+                    // return the new emulator id for the started emulators
+                    return emulatorId;
+                });
+            } else {
+                // We timed out waiting for the boot to happen
+                return null;
+            }
         });
     });
 };
 
 /*
- * Waits for an emulator with given uuid to apear on the started-emulator list.
- * Returns a promise with this emulator's ID.
+ * Waits for an emulator to boot on a given port.
+ * Returns this emulator's ID in a promise.
  */
-module.exports.wait_for_emulator = function(uuid) {
+module.exports.wait_for_emulator = function (port) {
     var self = this;
-    return self.list_started()
-    .then(function(new_started) {
-        var emulator_id = null;
-        var promises = [];
-
-        new_started.forEach(function (emulator) {
-            promises.push(
-                Adb.shell(emulator, 'getprop emu.uuid')
-                .then(function (output) {
-                    if (output.indexOf(uuid) >= 0) {
-                        emulator_id = emulator;
-                    }
-                })
-            );
+    return Q().then(function () {
+        var emulator_id = 'emulator-' + port;
+        return Adb.shell(emulator_id, 'getprop dev.bootcomplete').then(function (output) {
+            if (output.indexOf('1') >= 0) {
+                return emulator_id;
+            }
+            return self.wait_for_emulator(port);
+        }, function (error) {
+            if ((error && error.message &&
+            (error.message.indexOf('not found') > -1)) ||
+            (error.message.indexOf('device offline') > -1)) {
+                // emulator not yet started, continue waiting
+                return self.wait_for_emulator(port);
+            } else {
+                // something unexpected has happened
+                throw error;
+            }
         });
-
-        return Q.all(promises).then(function () {
-            return emulator_id || self.wait_for_emulator(uuid);
-        });
-     });
+    });
 };
 
 /*
- * Waits for the core android process of the emulator to start
+ * Waits for the core android process of the emulator to start. Returns a
+ * promise that resolves to a boolean indicating success. Not specifying a
+ * time_remaining or passing a negative value will cause it to wait forever
  */
-module.exports.wait_for_boot = function(emulator_id) {
+module.exports.wait_for_boot = function (emulator_id, time_remaining) {
     var self = this;
-    return Adb.shell(emulator_id, 'ps')
-    .then(function(output) {
+    return Adb.shell(emulator_id, 'ps').then(function (output) {
         if (output.match(/android\.process\.acore/)) {
-            return;
+            return true;
+        } else if (time_remaining === 0) {
+            return false;
         } else {
             process.stdout.write('.');
-            return Q.delay(3000).then(function() {
-                return self.wait_for_boot(emulator_id);
+
+            // Check at regular intervals
+            return Q.delay(time_remaining < CHECK_BOOTED_INTERVAL ? time_remaining : CHECK_BOOTED_INTERVAL).then(function () {
+                var updated_time = time_remaining >= 0 ? Math.max(time_remaining - CHECK_BOOTED_INTERVAL, 0) : time_remaining;
+                return self.wait_for_boot(emulator_id, updated_time);
             });
         }
     });
@@ -249,35 +382,33 @@ module.exports.wait_for_boot = function(emulator_id) {
  * TODO : Enter the stdin input required to complete the creation of an avd.
  * Returns a promise.
  */
-module.exports.create_image = function(name, target) {
-    console.log('Creating avd named ' + name);
+module.exports.create_image = function (name, target) {
+    console.log('Creating new avd named ' + name);
     if (target) {
-        return spawn('android', ['create', 'avd', '--name', name, '--target', target])
-        .then(null, function(error) {
+        return superspawn.spawn('android', ['create', 'avd', '--name', name, '--target', target]).then(null, function (error) {
             console.error('ERROR : Failed to create emulator image : ');
             console.error(' Do you have the latest android targets including ' + target + '?');
             console.error(error);
         });
     } else {
         console.log('WARNING : Project target not found, creating avd with a different target but the project may fail to install.');
-        return spawn('android', ['create', 'avd', '--name', name, '--target', this.list_targets()[0]])
-        .then(function() {
+        // TODO: there's a more robust method for finding targets in android_sdk.js
+        return superspawn.spawn('android', ['create', 'avd', '--name', name, '--target', this.list_targets()[0]]).then(function () {
             // TODO: This seems like another error case, even though it always happens.
             console.error('ERROR : Unable to create an avd emulator, no targets found.');
-            console.error('Please insure you have targets available by running the "android" command');
+            console.error('Ensure you have targets available by running the "android" command');
             return Q.reject();
-        }, function(error) {
+        }, function (error) {
             console.error('ERROR : Failed to create emulator image : ');
             console.error(error);
         });
     }
 };
 
-module.exports.resolveTarget = function(target) {
-    return this.list_started()
-    .then(function(emulator_list) {
+module.exports.resolveTarget = function (target) {
+    return this.list_started().then(function (emulator_list) {
         if (emulator_list.length < 1) {
-            return Q.reject('No started emulators found, please start an emultor before deploying your project.');
+            return Q.reject('No running Android emulators found, please start an emulator before deploying your project.');
         }
 
         // default emulator
@@ -286,9 +417,8 @@ module.exports.resolveTarget = function(target) {
             return Q.reject('Unable to find target \'' + target + '\'. Failed to deploy to emulator.');
         }
 
-        return build.detectArchitecture(target)
-        .then(function(arch) {
-            return {target:target, arch:arch, isEmulator:true};
+        return build.detectArchitecture(target).then(function (arch) {
+            return {target: target, arch: arch, isEmulator: true};
         });
     });
 };
@@ -299,7 +429,7 @@ module.exports.resolveTarget = function(target) {
  * If no started emulators are found, error out.
  * Returns a promise.
  */
-module.exports.install = function(givenTarget, buildResults) {
+module.exports.install = function (givenTarget, buildResults) {
 
     var target;
     var manifest = new AndroidManifest(path.join(__dirname, '../../AndroidManifest.xml'));
@@ -307,7 +437,7 @@ module.exports.install = function(givenTarget, buildResults) {
 
     // resolve the target emulator
     return Q().then(function () {
-        if (givenTarget && typeof givenTarget == 'object') {
+        if (givenTarget && typeof givenTarget === 'object') {
             return givenTarget;
         } else {
             return module.exports.resolveTarget(givenTarget);
@@ -321,41 +451,65 @@ module.exports.install = function(givenTarget, buildResults) {
     }).then(function () {
         // This promise is always resolved, even if 'adb uninstall' fails to uninstall app
         // or the app doesn't installed at all, so no error catching needed.
-        return Adb.uninstall(target.target, pkgName)
-        .then(function() {
+        return Q.when().then(function () {
 
             var apk_path = build.findBestApkForArchitecture(buildResults, target.arch);
             var execOptions = {
                 cwd: os.tmpdir(),
-                timeout:    INSTALL_COMMAND_TIMEOUT, // in milliseconds
+                timeout: INSTALL_COMMAND_TIMEOUT, // in milliseconds
                 killSignal: EXEC_KILL_SIGNAL
             };
 
             events.emit('log', 'Using apk: ' + apk_path);
+            events.emit('log', 'Package name: ' + pkgName);
             events.emit('verbose', 'Installing app on emulator...');
 
-            function exec(command, opts) {
+            // A special function to call adb install in specific environment w/ specific options.
+            // Introduced as a part of fix for http://issues.apache.org/jira/browse/CB-9119
+            // to workaround sporadic emulator hangs
+            function adbInstallWithOptions (target, apk, opts) {
+                events.emit('verbose', 'Installing apk ' + apk + ' on ' + target + '...');
+
+                var command = 'adb -s ' + target + ' install -r "' + apk + '"';
                 return Q.promise(function (resolve, reject) {
-                    child_process.exec(command, opts, function(err, stdout, stderr) {
+                    child_process.exec(command, opts, function (err, stdout, stderr) {
                         if (err) reject(new CordovaError('Error executing "' + command + '": ' + stderr));
-                        else resolve(stdout);
+                        // adb does not return an error code even if installation fails. Instead it puts a specific
+                        // message to stdout, so we have to use RegExp matching to detect installation failure.
+                        else if (/Failure/.test(stdout)) {
+                            if (stdout.match(/INSTALL_PARSE_FAILED_NO_CERTIFICATES/)) {
+                                stdout += 'Sign the build using \'-- --keystore\' or \'--buildConfig\'' +
+                                    ' or sign and deploy the unsigned apk manually using Android tools.';
+                            } else if (stdout.match(/INSTALL_FAILED_VERSION_DOWNGRADE/)) {
+                                stdout += 'You\'re trying to install apk with a lower versionCode that is already installed.' +
+                                    '\nEither uninstall an app or increment the versionCode.';
+                            }
+
+                            reject(new CordovaError('Failed to install apk to emulator: ' + stdout));
+                        } else resolve(stdout);
                     });
                 });
             }
 
-            var retriedInstall = retry.retryPromise(
-                NUM_INSTALL_RETRIES,
-                exec, 'adb -s ' + target.target + ' install -r "' + apk_path + '"', execOptions
-            );
+            function installPromise () {
+                return adbInstallWithOptions(target.target, apk_path, execOptions).catch(function (error) {
+                    // CB-9557 CB-10157 only uninstall and reinstall app if the one that
+                    // is already installed on device was signed w/different certificate
+                    if (!/INSTALL_PARSE_FAILED_INCONSISTENT_CERTIFICATES/.test(error.toString())) { throw error; }
 
-            return retriedInstall.then(function (output) {
-                if (output.match(/Failure/)) {
-                    return Q.reject(new CordovaError('Failed to install apk to emulator: ' + output));
-                } else {
-                    events.emit('log', 'INSTALL SUCCESS');
-                }
-            }, function (err) {
-                return Q.reject(new CordovaError('Failed to install apk to emulator: ' + err));
+                    events.emit('warn', 'Uninstalling app from device and reinstalling it because the ' +
+                        'currently installed app was signed with different key');
+
+                    // This promise is always resolved, even if 'adb uninstall' fails to uninstall app
+                    // or the app doesn't installed at all, so no error catching needed.
+                    return Adb.uninstall(target.target, pkgName).then(function () {
+                        return adbInstallWithOptions(target.target, apk_path, execOptions);
+                    });
+                });
+            }
+
+            return retry.retryPromise(NUM_INSTALL_RETRIES, installPromise).then(function (output) {
+                events.emit('log', 'INSTALL SUCCESS');
             });
         });
     // unlock screen
