@@ -1,293 +1,349 @@
 cordova.define('cordova/plugin_list', function(require, exports, module) {
 module.exports = [
-    {
-        "file": "plugins/org.nypr.cordova.wakeupplugin/www/wakeup.js",
-        "id": "org.nypr.cordova.wakeupplugin.Wakeup",
-        "clobbers": [
-            "wakeuptimer"
-        ]
-    },
-    {
-        "file": "plugins/com.phonegap.plugins.facebookconnect/facebookConnectPlugin.js",
-        "id": "com.phonegap.plugins.facebookconnect.FacebookConnectPlugin",
-        "clobbers": [
-            "facebookConnectPlugin"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file-transfer/www/FileTransferError.js",
-        "id": "cordova-plugin-file-transfer.FileTransferError",
-        "clobbers": [
-            "window.FileTransferError"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file-transfer/www/FileTransfer.js",
-        "id": "cordova-plugin-file-transfer.FileTransfer",
-        "clobbers": [
-            "window.FileTransfer"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-geolocation/www/android/geolocation.js",
-        "id": "cordova-plugin-geolocation.geolocation",
-        "clobbers": [
-            "navigator.geolocation"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-geolocation/www/PositionError.js",
-        "id": "cordova-plugin-geolocation.PositionError",
-        "runs": true
-    },
-    {
-        "file": "plugins/cordova-plugin-dialogs/www/notification.js",
-        "id": "cordova-plugin-dialogs.notification",
-        "merges": [
-            "navigator.notification"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-dialogs/www/android/notification.js",
-        "id": "cordova-plugin-dialogs.notification_android",
-        "merges": [
-            "navigator.notification"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-vibration/www/vibration.js",
-        "id": "cordova-plugin-vibration.notification",
-        "merges": [
-            "navigator.notification",
-            "navigator"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/DirectoryEntry.js",
-        "id": "cordova-plugin-file.DirectoryEntry",
-        "clobbers": [
-            "window.DirectoryEntry"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/DirectoryReader.js",
-        "id": "cordova-plugin-file.DirectoryReader",
-        "clobbers": [
-            "window.DirectoryReader"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/Entry.js",
-        "id": "cordova-plugin-file.Entry",
-        "clobbers": [
-            "window.Entry"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/File.js",
-        "id": "cordova-plugin-file.File",
-        "clobbers": [
-            "window.File"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileEntry.js",
-        "id": "cordova-plugin-file.FileEntry",
-        "clobbers": [
-            "window.FileEntry"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileError.js",
-        "id": "cordova-plugin-file.FileError",
-        "clobbers": [
-            "window.FileError"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileReader.js",
-        "id": "cordova-plugin-file.FileReader",
-        "clobbers": [
-            "window.FileReader"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileSystem.js",
-        "id": "cordova-plugin-file.FileSystem",
-        "clobbers": [
-            "window.FileSystem"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileUploadOptions.js",
-        "id": "cordova-plugin-file.FileUploadOptions",
-        "clobbers": [
-            "window.FileUploadOptions"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileUploadResult.js",
-        "id": "cordova-plugin-file.FileUploadResult",
-        "clobbers": [
-            "window.FileUploadResult"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/FileWriter.js",
-        "id": "cordova-plugin-file.FileWriter",
-        "clobbers": [
-            "window.FileWriter"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/Flags.js",
-        "id": "cordova-plugin-file.Flags",
-        "clobbers": [
-            "window.Flags"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/LocalFileSystem.js",
-        "id": "cordova-plugin-file.LocalFileSystem",
-        "clobbers": [
-            "window.LocalFileSystem"
-        ],
-        "merges": [
-            "window"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/Metadata.js",
-        "id": "cordova-plugin-file.Metadata",
-        "clobbers": [
-            "window.Metadata"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/ProgressEvent.js",
-        "id": "cordova-plugin-file.ProgressEvent",
-        "clobbers": [
-            "window.ProgressEvent"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/fileSystems.js",
-        "id": "cordova-plugin-file.fileSystems"
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/requestFileSystem.js",
-        "id": "cordova-plugin-file.requestFileSystem",
-        "clobbers": [
-            "window.requestFileSystem"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/resolveLocalFileSystemURI.js",
-        "id": "cordova-plugin-file.resolveLocalFileSystemURI",
-        "merges": [
-            "window"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/browser/isChrome.js",
-        "id": "cordova-plugin-file.isChrome",
-        "runs": true
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/android/FileSystem.js",
-        "id": "cordova-plugin-file.androidFileSystem",
-        "merges": [
-            "FileSystem"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/fileSystems-roots.js",
-        "id": "cordova-plugin-file.fileSystems-roots",
-        "runs": true
-    },
-    {
-        "file": "plugins/cordova-plugin-file/www/fileSystemPaths.js",
-        "id": "cordova-plugin-file.fileSystemPaths",
-        "merges": [
-            "cordova"
-        ],
-        "runs": true
-    },
-    {
-        "file": "plugins/cordova-plugin-device-orientation/www/CompassError.js",
-        "id": "cordova-plugin-device-orientation.CompassError",
-        "clobbers": [
-            "CompassError"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-device-orientation/www/CompassHeading.js",
-        "id": "cordova-plugin-device-orientation.CompassHeading",
-        "clobbers": [
-            "CompassHeading"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-device-orientation/www/compass.js",
-        "id": "cordova-plugin-device-orientation.compass",
-        "clobbers": [
-            "navigator.compass"
-        ]
-    },
-    {
-        "file": "plugins/com.jetboystudio.pebble.PebblePGPlugin/www/Pebble.js",
-        "id": "com.jetboystudio.pebble.PebblePGPlugin.Pebble",
-        "clobbers": [
-            "Pebble"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-network-information/www/network.js",
-        "id": "cordova-plugin-network-information.network",
-        "clobbers": [
-            "navigator.connection",
-            "navigator.network.connection"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-network-information/www/Connection.js",
-        "id": "cordova-plugin-network-information.Connection",
-        "clobbers": [
-            "Connection"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-device/www/device.js",
-        "id": "cordova-plugin-device.device",
-        "clobbers": [
-            "device"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-background-mode/www/background-mode.js",
-        "id": "cordova-plugin-background-mode.BackgroundMode",
-        "clobbers": [
-            "cordova.plugins.backgroundMode",
-            "plugin.backgroundMode"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-device-motion/www/Acceleration.js",
-        "id": "cordova-plugin-device-motion.Acceleration",
-        "clobbers": [
-            "Acceleration"
-        ]
-    },
-    {
-        "file": "plugins/cordova-plugin-device-motion/www/accelerometer.js",
-        "id": "cordova-plugin-device-motion.accelerometer",
-        "clobbers": [
-            "navigator.accelerometer"
-        ]
-    }
+  {
+    "id": "com.jetboystudio.pebble.PebblePGPlugin.Pebble",
+    "file": "plugins/com.jetboystudio.pebble.PebblePGPlugin/www/Pebble.js",
+    "pluginId": "com.jetboystudio.pebble.PebblePGPlugin",
+    "clobbers": [
+      "Pebble"
+    ]
+  },
+  {
+    "id": "com.phonegap.plugins.facebookconnect.FacebookConnectPlugin",
+    "file": "plugins/com.phonegap.plugins.facebookconnect/facebookConnectPlugin.js",
+    "pluginId": "com.phonegap.plugins.facebookconnect",
+    "clobbers": [
+      "facebookConnectPlugin"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device.device",
+    "file": "plugins/cordova-plugin-device/www/device.js",
+    "pluginId": "cordova-plugin-device",
+    "clobbers": [
+      "device"
+    ]
+  },
+  {
+    "id": "cordova-plugin-background-mode.BackgroundMode",
+    "file": "plugins/cordova-plugin-background-mode/www/background-mode.js",
+    "pluginId": "cordova-plugin-background-mode",
+    "clobbers": [
+      "cordova.plugins.backgroundMode",
+      "plugin.backgroundMode"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device-motion.Acceleration",
+    "file": "plugins/cordova-plugin-device-motion/www/Acceleration.js",
+    "pluginId": "cordova-plugin-device-motion",
+    "clobbers": [
+      "Acceleration"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device-motion.accelerometer",
+    "file": "plugins/cordova-plugin-device-motion/www/accelerometer.js",
+    "pluginId": "cordova-plugin-device-motion",
+    "clobbers": [
+      "navigator.accelerometer"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device-orientation.CompassError",
+    "file": "plugins/cordova-plugin-device-orientation/www/CompassError.js",
+    "pluginId": "cordova-plugin-device-orientation",
+    "clobbers": [
+      "CompassError"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device-orientation.CompassHeading",
+    "file": "plugins/cordova-plugin-device-orientation/www/CompassHeading.js",
+    "pluginId": "cordova-plugin-device-orientation",
+    "clobbers": [
+      "CompassHeading"
+    ]
+  },
+  {
+    "id": "cordova-plugin-device-orientation.compass",
+    "file": "plugins/cordova-plugin-device-orientation/www/compass.js",
+    "pluginId": "cordova-plugin-device-orientation",
+    "clobbers": [
+      "navigator.compass"
+    ]
+  },
+  {
+    "id": "cordova-plugin-dialogs.notification",
+    "file": "plugins/cordova-plugin-dialogs/www/notification.js",
+    "pluginId": "cordova-plugin-dialogs",
+    "merges": [
+      "navigator.notification"
+    ]
+  },
+  {
+    "id": "cordova-plugin-dialogs.notification_android",
+    "file": "plugins/cordova-plugin-dialogs/www/android/notification.js",
+    "pluginId": "cordova-plugin-dialogs",
+    "merges": [
+      "navigator.notification"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.DirectoryEntry",
+    "file": "plugins/cordova-plugin-file/www/DirectoryEntry.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.DirectoryEntry"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.DirectoryReader",
+    "file": "plugins/cordova-plugin-file/www/DirectoryReader.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.DirectoryReader"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.Entry",
+    "file": "plugins/cordova-plugin-file/www/Entry.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.Entry"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.File",
+    "file": "plugins/cordova-plugin-file/www/File.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.File"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileEntry",
+    "file": "plugins/cordova-plugin-file/www/FileEntry.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileEntry"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileError",
+    "file": "plugins/cordova-plugin-file/www/FileError.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileError"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileReader",
+    "file": "plugins/cordova-plugin-file/www/FileReader.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileReader"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileSystem",
+    "file": "plugins/cordova-plugin-file/www/FileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileSystem"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileUploadOptions",
+    "file": "plugins/cordova-plugin-file/www/FileUploadOptions.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileUploadOptions"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileUploadResult",
+    "file": "plugins/cordova-plugin-file/www/FileUploadResult.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileUploadResult"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.FileWriter",
+    "file": "plugins/cordova-plugin-file/www/FileWriter.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.FileWriter"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.Flags",
+    "file": "plugins/cordova-plugin-file/www/Flags.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.Flags"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.LocalFileSystem",
+    "file": "plugins/cordova-plugin-file/www/LocalFileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.LocalFileSystem"
+    ],
+    "merges": [
+      "window"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.Metadata",
+    "file": "plugins/cordova-plugin-file/www/Metadata.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.Metadata"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.ProgressEvent",
+    "file": "plugins/cordova-plugin-file/www/ProgressEvent.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.ProgressEvent"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.fileSystems",
+    "file": "plugins/cordova-plugin-file/www/fileSystems.js",
+    "pluginId": "cordova-plugin-file"
+  },
+  {
+    "id": "cordova-plugin-file.requestFileSystem",
+    "file": "plugins/cordova-plugin-file/www/requestFileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "clobbers": [
+      "window.requestFileSystem"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.resolveLocalFileSystemURI",
+    "file": "plugins/cordova-plugin-file/www/resolveLocalFileSystemURI.js",
+    "pluginId": "cordova-plugin-file",
+    "merges": [
+      "window"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.isChrome",
+    "file": "plugins/cordova-plugin-file/www/browser/isChrome.js",
+    "pluginId": "cordova-plugin-file",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-file.androidFileSystem",
+    "file": "plugins/cordova-plugin-file/www/android/FileSystem.js",
+    "pluginId": "cordova-plugin-file",
+    "merges": [
+      "FileSystem"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file.fileSystems-roots",
+    "file": "plugins/cordova-plugin-file/www/fileSystems-roots.js",
+    "pluginId": "cordova-plugin-file",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-file.fileSystemPaths",
+    "file": "plugins/cordova-plugin-file/www/fileSystemPaths.js",
+    "pluginId": "cordova-plugin-file",
+    "merges": [
+      "cordova"
+    ],
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-file-transfer.FileTransferError",
+    "file": "plugins/cordova-plugin-file-transfer/www/FileTransferError.js",
+    "pluginId": "cordova-plugin-file-transfer",
+    "clobbers": [
+      "window.FileTransferError"
+    ]
+  },
+  {
+    "id": "cordova-plugin-file-transfer.FileTransfer",
+    "file": "plugins/cordova-plugin-file-transfer/www/FileTransfer.js",
+    "pluginId": "cordova-plugin-file-transfer",
+    "clobbers": [
+      "window.FileTransfer"
+    ]
+  },
+  {
+    "id": "cordova-plugin-geolocation.geolocation",
+    "file": "plugins/cordova-plugin-geolocation/www/android/geolocation.js",
+    "pluginId": "cordova-plugin-geolocation",
+    "clobbers": [
+      "navigator.geolocation"
+    ]
+  },
+  {
+    "id": "cordova-plugin-geolocation.PositionError",
+    "file": "plugins/cordova-plugin-geolocation/www/PositionError.js",
+    "pluginId": "cordova-plugin-geolocation",
+    "runs": true
+  },
+  {
+    "id": "cordova-plugin-network-information.network",
+    "file": "plugins/cordova-plugin-network-information/www/network.js",
+    "pluginId": "cordova-plugin-network-information",
+    "clobbers": [
+      "navigator.connection",
+      "navigator.network.connection"
+    ]
+  },
+  {
+    "id": "cordova-plugin-network-information.Connection",
+    "file": "plugins/cordova-plugin-network-information/www/Connection.js",
+    "pluginId": "cordova-plugin-network-information",
+    "clobbers": [
+      "Connection"
+    ]
+  },
+  {
+    "id": "cordova-plugin-vibration.notification",
+    "file": "plugins/cordova-plugin-vibration/www/vibration.js",
+    "pluginId": "cordova-plugin-vibration",
+    "merges": [
+      "navigator.notification",
+      "navigator"
+    ]
+  },
+  {
+    "id": "org.nypr.cordova.wakeupplugin.Wakeup",
+    "file": "plugins/org.nypr.cordova.wakeupplugin/www/wakeup.js",
+    "pluginId": "org.nypr.cordova.wakeupplugin",
+    "clobbers": [
+      "wakeuptimer"
+    ]
+  }
 ];
 module.exports.metadata = 
 // TOP OF METADATA
-{}
+{
+  "com.jetboystudio.pebble.PebblePGPlugin": "0.2.6",
+  "com.phonegap.plugins.facebookconnect": "0.11.0",
+  "cordova-plugin-device": "1.1.6",
+  "cordova-plugin-background-mode": "0.7.2",
+  "cordova-plugin-device-motion": "1.2.5",
+  "cordova-plugin-device-orientation": "1.0.3",
+  "cordova-plugin-dialogs": "1.2.1",
+  "cordova-plugin-file": "4.2.0",
+  "cordova-plugin-file-transfer": "1.5.1",
+  "cordova-plugin-geolocation": "2.2.0",
+  "cordova-plugin-network-information": "1.3.2",
+  "cordova-plugin-vibration": "2.1.1",
+  "cordova-plugin-whitelist": "1.2.2",
+  "org.nypr.cordova.wakeupplugin": "0.1.0"
+};
 // BOTTOM OF METADATA
 });
