@@ -19,21 +19,20 @@
 
 package ws.wigo.geotrail;
 
-import android.annotation.TargetApi;
 import android.os.Bundle;
 import org.apache.cordova.*;
 
-@TargetApi(19)
 public class MainActivity extends CordovaActivity
 {
     @Override
     public void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
-        //20160507 if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.KITKAT)
-        if (android.os.Build.VERSION.SDK_INT >= 18) // KITKAT is 19
-        {
-            android.webkit.WebView.setWebContentsDebuggingEnabled(true);
+
+        // enable Cordova apps to be started in the background
+        Bundle extras = getIntent().getExtras();
+        if (extras != null && extras.getBoolean("cdvStartInBackground", false)) {
+            moveTaskToBack(true);
         }
 
         // Set by <content src="index.html" /> in config.xml
